bitcoin-ingestor/pyproject.toml
[project]
name = "market-data-ingestion"
version = "0.1.0"
description = "Real-time Binance market data ingestion to S3"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.34.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "moto>=4.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
